refactor: code refactoring
