+++ src/routes/graphql/schemas/member-types.ts
@@ -1,0 +1,22 @@
+import {
+  GraphQLEnumType,
+  GraphQLFloat,
+  GraphQLInt,
+  GraphQLObjectType,
+} from 'graphql/index.js';
+
+export const MemberTypeId = new GraphQLEnumType({
+  name: 'MemberTypeId',
+  values: {
+    basic: { value: 'basic' },
+    business: { value: 'business' },
+  },
+});
+export const MemberTypes = new GraphQLObjectType({
+  name: 'memberTypes',
+  fields: {
+    id: { type: MemberTypeId },
+    discount: { type: GraphQLFloat },
+    postsLimitPerMonth: { type: GraphQLInt },
+  },
+});
