+++ src/routes/graphql/schemas/posts.ts
@@ -1,0 +1,34 @@
+import {
+  GraphQLInputObjectType,
+  GraphQLObjectType,
+  GraphQLString,
+} from 'graphql/index.js';
+import { UUIDType } from '../types/uuid.js';
+
+export const PostTypes = new GraphQLObjectType({
+  name: 'posts',
+  fields: {
+    id: { type: UUIDType },
+    title: { type: GraphQLString },
+    content: { type: GraphQLString },
+    authorId: { type: UUIDType },
+  },
+});
+
+export const CreatePostInputType = new GraphQLInputObjectType({
+  name: 'CreatePostInput',
+  fields: {
+    title: { type: GraphQLString },
+    content: { type: GraphQLString },
+    authorId: { type: UUIDType },
+  },
+});
+
+export const ChangePostInputType = new GraphQLInputObjectType({
+  name: 'ChangePostInput',
+  fields: {
+    title: { type: GraphQLString },
+    content: { type: GraphQLString },
+    authorId: { type: UUIDType },
+  },
+});
